% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.seqre_result}
\alias{autoplot.seqre_result}
\title{Create a SeqRe Result ggplot}
\usage{
autoplot.seqre_result(x, ..., func = "density", confint = NULL)
}
\arguments{
\item{x}{A \code{seqre_result} object.}

\item{...}{Additional arguments discarded from \code{plot}.}

\item{func}{Either 'distribution', 'density', or 'gradient'.}

\item{confint}{A decimal value indicating the confidence interval width (e.g.
0.95 for a 95\% confidence interval). Defaults to \code{NULL}, in which case no
confidence intervals will be drawn.}
}
\value{
A \code{ggplot} object.
}
\description{
Create a SeqRe Result ggplot
}
