% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{functional}
\alias{functional}
\title{Obtain Functionals from a CopRe Result}
\usage{
functional(obj, f, ..., mean = FALSE)
}
\arguments{
\item{obj}{A \code{copre_result} object.}

\item{f}{A list of functions.}

\item{...}{Additional arguments passed to \code{f}.}

\item{mean}{A logical value indicating whether or not to obtain the
functional from the pointwise mean of the sampled distributions or from each
individually.}
}
\value{
The integral over the \code{copre_result} grid of the functions in the
list multiplied by the density of each sample distribution in \code{obj}.
}
\description{
Obtain Functionals from a CopRe Result
}
