% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_measure.R
\name{G_normgam}
\alias{G_normgam}
\title{A Normal-Inverse-Gamma Base Measure for Normal Mixture Models.}
\usage{
G_normgam(mu, tau, s, S)
}
\arguments{
\item{mu}{The mean parameter.}

\item{tau}{The variance scaling parameter.}

\item{s}{The primary shape parameter for the Inverse-Gamma component.}

\item{S}{The secondary shape parameter for the Inverse-Gamma component.}
}
\value{
A \code{base_measure} object for use in the exchangeable sequence
 resampling scheme for mixtures.
}
\description{
A Normal-Inverse-Gamma Base Measure for Normal Mixture Models.
}
\seealso{
{[base_measure(), seqre()]}
}
