% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_measure.R
\name{G_normls}
\alias{G_normls}
\title{Normal-Inverse-Gamma Base Measure for Location-Scale Normal Mixture Models.}
\usage{
G_normls(mu = 0, tau = 1, s = 1, S = 1, a = NULL, A = NULL, w = NULL, W = NULL)
}
\arguments{
\item{mu}{The mean parameter.}

\item{tau}{The variance scaling parameter.}

\item{s}{The primary shape parameter for the Inverse-Gamma component.}

\item{S}{The secondary shape parameter for the Inverse-Gamma component.}

\item{a}{The prior mean parameter for \code{mu}.}

\item{A}{The prior variance for \code{mu}.}

\item{w}{The prior primary shape parameter for \code{tau}.}

\item{W}{The prior secondary shape parameter for \code{tau}.}
}
\value{
A \code{base_measure} object for use in the sequence resampling scheme for
mixtures.
}
\description{
Normal-Inverse-Gamma Base Measure for Location-Scale Normal Mixture Models.
}
\seealso{
\code{\link[=base_measure]{base_measure()}}, \code{\link[=seqre]{seqre()}}
}
