% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{length.grideval_result}
\alias{length.grideval_result}
\title{Length}
\usage{
\method{length}{grideval_result}(x)
}
\arguments{
\item{x}{A \code{grideval_result} object.}
}
\value{
The number of samples \code{k} in \code{obj}.
}
\description{
Length
}
