% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{moments}
\alias{moments}
\alias{moments.seqre_result}
\alias{moments.grideval_result}
\title{Obtain Moments from a CopRe or SeqRe Result}
\usage{
moments(obj, mom, cntrl = TRUE, grd = NULL)

\method{moments}{seqre_result}(obj, mom, cntrl = TRUE, grd = NULL)

\method{moments}{grideval_result}(obj, mom, cntrl = TRUE, grd = NULL)
}
\arguments{
\item{obj}{A \code{copre_result} or \code{seqre_result} object.}

\item{mom}{A numeric scalar indicating the moment to calculate.}

\item{cntrl}{A logical value indicating whether the moment should be central
or not. Defaults to \code{TRUE}.}

\item{grd}{A numeric vector of grid values on which the density function
samples in \code{obj} should be calculated for trapezoidal integration.}
}
\value{
A vector of moment values for each sampled distribution in \code{obj}.
}
\description{
Obtain Moments from a CopRe or SeqRe Result
}
\section{Methods (by class)}{
\itemize{
\item \code{moments(seqre_result)}: Moment calculation method for \code{seqre_result} objects.

\item \code{moments(grideval_result)}: Moment calculation method for \code{grideval_result} objects.

}}
