% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.copre_result}
\alias{plot.copre_result}
\title{CopRe Result Plotter}
\usage{
\method{plot}{copre_result}(x, ..., func = "density", confint = NULL)
}
\arguments{
\item{x}{A \code{copre_result} object.}

\item{...}{Additional arguments discarded from \code{plot}.}

\item{func}{Either 'distribution', 'density', or 'gradient'.}

\item{confint}{A decimal value indicating the confidence interval width (e.g.
0.95 for a 95\% confidence interval). Defaults to \code{NULL}, in which case
no confidence intervals will be drawn.}
}
\value{
A \code{ggplot} object.
}
\description{
CopRe Result Plotter
}
