% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mdpolya_result}
\alias{plot.mdpolya_result}
\title{Plotting method for \code{mdpolya_result} objects}
\usage{
\method{plot}{mdpolya_result}(x, ..., grd = NULL, func = "density", confint = NULL, nthreads = 1)
}
\arguments{
\item{x}{An \code{mdpolya_result} object, substituting \code{obj}.}

\item{...}{Additional arguments discarded from \code{plot}.}

\item{grd}{For \code{mdpolya_result} objects, a numeric vector of \code{m}
grid points.}

\item{func}{Either 'distribution', 'density', or 'gradient'.}

\item{confint}{A decimal value indicating the confidence interval width (e.g.
0.95 for a 95% confidence interval). Defaults to \code{NULL}, in which case
no confidence intervals will be drawn.}

\item{nthreads}{The number of parallel threads to launch with OpenMP.}
}
\value{
A \code{ggplot} object.
}
\description{
Plotting method for \code{mdpolya_result} objects
}
