% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{register_autoplot_s3_methods}
\alias{register_autoplot_s3_methods}
\title{Register \code{autoplot} methods to \code{ggplot2}}
\usage{
register_autoplot_s3_methods()
}
\value{
None
}
\description{
Register \code{autoplot} methods to \code{ggplot2}
}
