% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_s3.R
\name{register_s3_method}
\alias{register_s3_method}
\title{Register S3 Methods from External Packages}
\usage{
register_s3_method(pkg, generic, class, fun = NULL)
}
\arguments{
\item{pkg}{Package name.}

\item{generic}{Generic function name.}

\item{class}{Class name.}

\item{fun}{Optional custum function name.}
}
\value{
None
}
\description{
https://github.com/tidyverse/hms/blob/master/R/zzz.R
}
