% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_measure.R
\name{Sq_gnedin0}
\alias{Sq_gnedin0}
\title{Collapsed Gnedin Process Sequence Measure.}
\usage{
Sq_gnedin0(gamma)
}
\arguments{
\item{gamma}{The gamma parameter for the Gnedin process with xi set to 0.
Bounded to \code{[0, 1]}.}
}
\value{
A \code{seq_measure} object for use in the exchangeable sequence
 resampling scheme for mixtures.
}
\description{
Collapsed Gnedin Process Sequence Measure.
}
\seealso{
{[seq_measure(), exseqre()]}
}
