% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_measure.R
\name{base_measure}
\alias{base_measure}
\title{Base Measure for Mixture Models}
\usage{
base_measure(pars, G, q0, q, eval)
}
\arguments{
\item{pars}{A list of hyperparameters used in \code{G}.}

\item{G}{A function which generates mixture component parameters from the
conditional posterior given data \code{y}, or generates from the prior if
\code{y} is empty or \code{NULL}.}

\item{q0}{A marginal likelihood function on \code{y} using the values in
\code{pars} (hard-coded, not as an argument).}

\item{q}{A likelihood function on \code{y} given \code{theta}, a mixture
component parameter drawn from \code{G}.}

\item{eval}{An evaluation function taking \code{phi}, a list of mixture,
matrices as in \code{mdp()}, \code{grd}, a grid vector, \code{f}, a
character string indicating whether to calculate the gradient, density, or
distribution function, and \code{nthreads}, a number of threads to utilize
for parallel execution.}
}
\value{
A \code{base_measure} object for use in the exchangeable sequence
 resampling scheme for mixtures.
}
\description{
A structure for wrapping base measures as in Escobar and West
(1995).
}
\references{
\itemize{
 \item Escobar M. D., West, M. (1995) Bayesian Density Estimation and
 Inference Using Mixtures. Journal of the American Statistical Association.
 DOI: \doi{10.1080/01621459.1995.10476550}
 }
 @export
}
\seealso{
{[exseqre()]}
}
