% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqre.R
\name{seqre}
\alias{seqre}
\title{Sequence Resampling}
\usage{
seqre(data, k, s_msr, b_msr, marg = FALSE, burn = 1000, thin = 150)
}
\arguments{
\item{data}{The data from which to sample predictive distributions.}

\item{k}{The number of predictive samples to draw.}

\item{s_msr}{A \code{seq_measure} object.}

\item{b_msr}{A univariate \code{base_measure} object.}

\item{burn}{The number of initial sampling iterations to discard, will be
truncated if a non-integer.}

\item{thin}{The number of sampling iterations to discard between records,
will be truncated if a non-integer.}
}
\value{
A \code{unimplemented} object.
}
\description{
A function that samples predictive distributions for univariate
 continuous data using exchangeable predictive extension.
}
\seealso{
[seq_measure()], [base_measure()]
}
