% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R, R/seqre.R
\name{[[.seqreresult}
\alias{[[.seqreresult}
\alias{seqre}
\title{Sequence Resampling}
\usage{
\method{[[}{seqreresult}(obj, i)

seqre(obj, inc = 1000, eps = 0.001, max_it = 100)
}
\arguments{
\item{obj}{A \code{seqre_result} object, usually output from \code{gibbsmix()}.}

\item{i}{A numeric vector of sample indices.}

\item{inc}{A positive integer increment value for the number of predictive
samples to take each convergence check.}

\item{eps}{An error value which determines the convergence approximation.}

\item{max_it}{A positive integer maximum number of iterations before halting.}
}
\value{
A \code{seqre_result} object, or a list of two \code{seqre_result} objects if
\code{keep_marg} is \code{TRUE}.
}
\description{
A function that samples predictive distributions for univariate
continuous data using exchangeable predictive extension.
}
\section{Functions}{
\itemize{
\item \code{[[}: Subset method for \code{seqre_result} objects

}}
\seealso{
\code{\link[=gibbsmix]{gibbsmix()}}
}
